# perf: read counters with RDPMC when possible on amd64