--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,5 @@
 module github.com/aclements/go-perfevent
 
-go 1.21
+go 1.22
 
-require golang.org/x/sys v0.17.0
+require golang.org/x/sys v0.27.0