--- a/perf/rdpmc_amd64.go
+++ b/perf/rdpmc_amd64.go
@@ -1,0 +1,7 @@
+// Copyright 2024 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package perf
+
+func rdpmc(counter uint32) int64